폴더구성 변경 및 route 변경 완료
