+++ src/features/calendar/Calendar.tsx
@@ -1,15 +1,16 @@
+import { cn } from "@/lib/utils"
 import CalendarSideBar from '@/components/calendar/SideBar';
 import CalendarContent from '@/components/calendar';
 
-import '@/view/calendar/calendar.scss';
-
-const Calendar: React.FC = () => {
+export default function Calendar() {
   return (
-    <div className='calendar'>
+    <div
+      className={cn(
+        "flex w-full h-full p-[10px]" 
+      )}
+    >
       <CalendarSideBar />
       <CalendarContent />
     </div>
   );
 }
-
-export default Calendar;
