+++ src/features/culture/Culture.tsx
@@ -1,0 +1,13 @@
+import { cn } from "@/lib/utils"
+
+export default function Culture() {
+  return (
+    <div
+      className={cn(
+        "flex w-full h-full p-[10px]" 
+      )}
+    >
+      <h1>culture page</h1>
+    </div>
+  );
+}
