+++ src/features/dashboard/Dashboard.tsx
@@ -1,0 +1,13 @@
+import { cn } from "@/lib/utils"
+
+export default function Dashboard() {
+  return (
+    <div
+      className={cn(
+        "flex w-full h-full p-[10px]" 
+      )}
+    >
+      <h1>dashboard page</h1>
+    </div>
+  );
+}
