+++ src/features/sidebar/nav-content.tsx
@@ -21,7 +21,8 @@ export function NavContent({
   items: {
     title: string
     url: string
-    icon?: LucideIcon
+    // icon?: LucideIcon
+    icon?: any
     isActive?: boolean
     items?: {
       title: string
