+++ src/features/work/Work.tsx
@@ -1,0 +1,13 @@
+import { cn } from "@/lib/utils"
+
+export default function Work() {
+  return (
+    <div
+      className={cn(
+        "flex w-full h-full p-[10px]" 
+      )}
+    >
+      <h1>work page</h1>
+    </div>
+  );
+}
