--- src/hooks/useMobile.tsx
@@ -1,20 +1,0 @@
-import { useEffect, useState } from "react"
-
-export const useIsMobile = () => {
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const checkDevice = () => {
-      setIsMobile(window.innerWidth < 1024); // lg 브레이크포인트
-    }
-
-    checkDevice();
-    window.addEventListener('resize', checkDevice);
-
-    return () => {
-      window.removeEventListener('resize', checkDevice);
-    }
-  }, []);
-
-  return isMobile;
-};
