--- src/view/calendar/calendar.scss
@@ -1,12 +1,0 @@
-@use '@/assets/scss/variables' as var;
-
-.calendar {
-    display: flex;
-
-    width: 100%;
-    padding: 15px;
-    margin: 70px 15px 15px 15px;
-
-    box-shadow: var.$box-shadow;
-    border-radius: var.$box-radius;
-}
