--- src/view/culture/Culture.tsx
@@ -1,11 +1,0 @@
-import '@/view/culture/culture.scss';
-
-const Culture: React.FC = () => {
-  return (
-    <div className='culture'>
-      <h1>culture page</h1>
-    </div>
-  );
-}
-
-export default Culture;
