--- src/view/dashboard/Dashboard.tsx
@@ -1,11 +1,0 @@
-import '@/view/dashboard/dashboard.scss';
-
-const Dashboard: React.FC = () => {
-  return (
-    <div className='home'>
-      Dashboard
-    </div>
-  );
-}
-
-export default Dashboard;
