--- src/view/dashboard/dashboard.scss
@@ -1,10 +1,0 @@
-@use '@/assets/scss/variables' as var;
-
-.home {
-    width: 100%;
-    padding: 15px;
-    margin: 70px 15px 15px 15px;
-
-    box-shadow: var.$box-shadow;
-    border-radius: var.$box-radius;
-}
