--- src/view/layout/Layout.tsx
@@ -1,17 +1,0 @@
-import { Outlet } from 'react-router-dom';
-import Nav from '@/components/nav/Nav';
-import Header from '@/components/header/Header';
-
-import '@/view/layout/layout.scss';
-
-const Layout: React.FC = () => {
-  return (
-    <div className='layout'>
-      <Header />
-      <Nav />
-      <Outlet />
-    </div>
-  );
-}
-
-export default Layout;
