--- src/view/layout/layout.scss
@@ -1,6 +1,0 @@
-.layout {
-    display: flex;
-    position: relative;
-    height: 100vh;
-    width: 100vw;
-}
