--- src/view/work/Work.tsx
@@ -1,11 +1,0 @@
-import '@/view/work/work.scss';
-
-const Work: React.FC = () => {
-  return (
-    <div className='work'>
-      <h1>work page</h1>
-    </div>
-  );
-}
-
-export default Work;
